Fixed buffer overflow vulnerability
Fixed buffer overflow vulnerability in adapter.c,
where the code was failing to check the size of a registry
string read by NdisReadConfiguration before copying it
to a fixed length buffer.

The vulnerability could potentially allow arbitrary
code execution in the kernel context of a signed driver,
however it requires local Admin privileges to exploit.

https://github.com/Rootkitsmm/OpenVpn-Pool-Overflow/blob/master/README.md

Signed-off-by: James Yonan <james@openvpn.net>