Fix possible buffer overread in HandleIPv6NeighborDiscovery()
Add missing length check for "is the packet big enough to contain a full
ICMPv6 header" to HandleIPv6NeighborDiscovery().

This bug can be triggered by a privileged local user sending a malformed
IPv6 packet over a raw socket to a link-local address on the tun interface.
The effect is a buffer overread by up to 24 bytes, which might have no
effects at all or cause a bluescreen, depending on kernel memory layout.
Memory contents overread will not be leaked.

If bridging to a LAN interface, injection of a malformed packet over
LAN->bridge->TAP adapter might be possible.  But if the TAP adapter is
used in "tap" mode (which is the only config that a bridge setup can
technically be used successfully), the problematic code is not executed.

Discovered and reported by Ludvig Strigeus <strigeus@gmail.com>.

v2: Typo in PacketLength vs. packetLength
v3: increase TAP_DRIVER_MINOR_VERSION to "0x04", update copyright year
v4: version.m4, bump version 9.21.2 -> 9.22.1
v5: add impact analysis

Signed-off-by: Gert Doering <gert@greenie.muc.de>
Tested-By: Gert Doering <gert@greenie.muc.de>

--
This is the same code as in v2, but driver versions have been bounced
to 9.22.1 / 4.4 to make the change more visible - most dialogues only
show "9.21" and to see the difference "9.21.2 -> 9.21.3" is not obvious.

Added impact analysis as discussed on IRC to avoid useless panic.  Updating
is good, drastic emergency activity is not needed.