Switch to VS 2017 makefiles
This is a straightforward translation of the Win7 DDK SOURCES file.

Note that TargetVersion=Windows7 sets NTDDI_VERSION correctly and
automatically picks the correct buffer overflow library for backwards
compatibility

The new toolchain does introduce a number of warnings that are benign
but should be addressed at a later date.