Give stampinf correct DriverVer information
VS2017 has some nice INF-processing infrastructure, but it uses stampinf
which insists on adding the DriverVer directive to the INF file itself.
This overwrites whatever the INF template produces.

This commit plumbs the INF template values in through stampinf. It also
uses the new and improved four-field version style instead of the old
a.00.00.b style.

Note that the VS2017 GUI field "Driver Version Number" is stored in the
project file field `TimeStamp`.