Use CurrentAddress for MAC instead of PermanentAddress
While checking packets for proxy arp etc., use of either the
current or permanent address worked in the past when we used
to overwrite the PermanentAddress with the user-defined value
(if any). Since commit f4f6464cac we directly set the
CurrentAddress to the customized value and leave the
PermanentAddress unchanged.

The DHCP code is unchanged as it already uses the CurrentAddress.

Also update the description of these fields in comments to better
match the code.

Signed-off-by: Selva Nair <selva.nair@gmail.com>