rxpath: fix broken TCP connections on Windows Server
Userspace OpenVPN implementation doesn't support
concurrent pending writes to tun. This is not an issue,
assuming that tun write is instantaneous (why would it not be?).

In short, this is what current driver implementation does on tun write:

 1. allocate NBL
 2. mark write requests as "pending"
 3. indicate NBL to NDIS
 4. when NDIS returns NBL ownership,
complete pending write request

On Windows Server 2019 and 2022 there are scenarios
where there is a noticeable (over second) second delay
between 3. and 4., and during that time no tun write requests
are performed. This breaks TCP communication inside tun, like "iperf3 -R".

I managed to break into debugger during that delay
and the current owner of NBL was TCP/IP protocol driver.
It is not clear for me why this delay in TCP/IP driver happens
and apparently only on Server editions.

More info: https://community.osr.com/discussion/293415/delay-in-calling-returnnetbufferlistshandler-on-windows-server-2022

As a workaround, use NDIS_RECEIVE_FLAGS_RESOURCES
flag with NdisMIndicateReceiveNetBufferLists call, which tells
NDIS to make a copy of NBL, which enables us to immediately free
NBL and complete request in place. Using this flag may have
performance implications according to MSFT, but driver's peformance
is already poor and according to my tests this change doesn't make it
worse. For the performance one should use dco-win.

Signed-off-by: Lev Stipakov <lev@openvpn.net>